* package upgrades * removed yarn.lock * renamed populateApidoc/Swagger scripts to generateX
