+++ package.json
@@ -17,10 +17,10 @@
     "vscolprod": "better-npm-run val-col-prod",
     "stability": "better-npm-run stability",
     "typeorm": "ts-node --files -r tsconfig-paths/register ./node_modules/.bin/typeorm",
-    "swagger": "ts-node --files -r tsconfig-paths/register src/scripts/populateSwagger.ts",
+    "swagger": "ts-node --files -r tsconfig-paths/register src/scripts/generateSwagger.ts",
     "mergeswagger": "ts-node --files -r tsconfig-paths/register src/scripts/mergeSwaggerFile.ts",
     "testurls": "ts-node --files -r tsconfig-paths/register src/scripts/testUrls.ts",
-    "apidoc": "ts-node --files -r tsconfig-paths/register src/scripts/populateApiDoc.ts"
+    "apidoc": "ts-node --files -r tsconfig-paths/register src/scripts/generateApiDoc.ts"
   },
   "betterScripts": {
     "api-dev": {
@@ -59,7 +59,7 @@
   },
   "dependencies": {
     "@koa/cors": "^3.0.0",
-    "@sentry/node": "^5.20.0",
+    "@sentry/node": "^5.20.1",
     "bech32": "^1.1.3",
     "bech32-buffer": "^0.1.2",
     "better-npm-run": "^0.1.1",
@@ -75,7 +75,7 @@
     "koa-add-trailing-slashes": "^2.0.1",
     "koa-body": "^4.1.0",
     "koa-helmet": "^5",
-    "koa-joi-controllers": "^1.2.2",
+    "koa-joi-controllers": "^1.2.3",
     "koa-morgan": "^1.0.1",
     "koa-mount": "^4.0.0",
     "koa-proxy": "^1.0.0-alpha.3",
@@ -103,7 +103,7 @@
   "devDependencies": {
     "@types/bluebird": "^3.5.31",
     "@types/got": "^9.6.9",
-    "@types/jest": "^26",
+    "@types/jest": "^26.0.8",
     "@types/koa": "^2.0.48",
     "@types/koa-router": "^7.0.40",
     "@types/lodash": "^4.14.158",
@@ -113,16 +113,16 @@
     "apidoc-core": "^0.11.1",
     "apidoc-swagger": "git+https://github.com/terra-project/apidoc-swagger.git",
     "husky": "^4.2.3",
-    "jest": "^26",
-    "jest-circus": "^26.0.1",
+    "jest": "^26.2.2",
+    "jest-circus": "^26.2.2",
     "jest-extended": "^0.11.5",
     "js-yaml": "^3.14.0",
     "lint-staged": "^10.0.9",
     "nodemon": "^2",
     "prettier": "^2",
     "supertest": "^4.0.2",
-    "ts-jest": "^26",
-    "tslint": "^6",
+    "ts-jest": "^26.1.4",
+    "tslint": "^6.1.3",
     "tslint-config-airbnb": "^5.11.2",
     "tslint-config-prettier": "^1.18.0",
     "yargs": "^15.3.1"
